requirements.txt
streamlit
requests
python-dotenv
openai>=1.0.0
